cleaned up and added code
if applied this commit will rearrange some code to be in better order. It will also add express code for the endpoint /movies/[title].

made table for endpoints

if applied this commit will make a table detailing endpoints for the api.

added express code

if applied this commit will add express code for the endpoint /movie/[genre]

added express code

if applied this commit will add express code for the endpoint /movies/[director]

added code to register and delete accounts

if applied this commit will add middleware that allows a person to register for an account as well as deleting that account.

added middleware and updated code

if applied this commit will add middleware for updating an accounts username, along with updating code to better reflect its purpose in the app.delete middleware

middleware added

if applied this commit will add middleware that allows the user to add a movie to their favorites

middleware added - remove

if applied this commit will add middleware that will remove a movie from the user's favorites list.

updated punctuation

if applied this commit will update the punctuation in the middleware for getting the director

updated errors in code

if applied this commit will update errors found in the middleware for title, genre, and director